internal/appsec/dyngo/instrumentation/httpinstr: renaming package into httpsec
